+++ driver/connection.go
@@ -183,6 +183,7 @@ type conn struct {
 
 	serverOptions p.Options[p.ConnectOption]
 	hdbVersion    *Version
+	hdbDfv        int
 
 	pr *p.Reader
 	pw *p.Writer
@@ -284,6 +285,7 @@ func initConn(ctx context.Context, metrics *metrics, attrs *connAttrs, auth *p.A
 	}
 
 	c.hdbVersion = parseVersion(c.versionString())
+	c.hdbDfv = int(c.serverOptions[p.CoDataFormatVersion2].(int32))
 
 	if attrs._defaultSchema != "" {
 		if _, err := c.ExecContext(ctx, strings.Join([]string{setDefaultSchema, Identifier(attrs._defaultSchema).String()}, " "), nil); err != nil {
@@ -1132,7 +1134,6 @@ func (c *conn) _authenticate(auth *p.Auth, applicationName string, dfv int, loca
 	if err != nil {
 		return 0, nil, err
 	}
-	//co := c.defaultClientOptions()
 
 	co := func() p.Options[p.ConnectOption] {
 		co := p.Options[p.ConnectOption]{
@@ -1163,9 +1164,6 @@ func (c *conn) _authenticate(auth *p.Auth, applicationName string, dfv int, loca
 			c.pr.Read(finalReply)
 		case p.PkConnectOptions:
 			c.pr.Read(&co)
-			// set data format version
-			// TODO generalize for sniffer
-			c.pr.SetDfv(int(co[p.CoDataFormatVersion2].(int32)))
 		}
 	}); err != nil {
 		return 0, nil, err
@@ -1182,7 +1180,7 @@ func (c *conn) _queryDirect(query string, commit bool) (driver.Rows, error) {
 	}
 
 	qr := &queryResult{conn: c}
-	meta := &p.ResultMetadata{}
+	meta := &p.ResultMetadata{Dfv: c.hdbDfv}
 	resSet := &p.Resultset{}
 
 	if err := c.pr.IterateParts(func(ph *p.PartHeader) {
@@ -1239,8 +1237,8 @@ func (c *conn) _prepare(query string) (*prepareResult, error) {
 	}
 
 	pr := &prepareResult{}
-	resMeta := &p.ResultMetadata{}
-	prmMeta := &p.ParameterMetadata{}
+	resMeta := &p.ResultMetadata{Dfv: c.hdbDfv}
+	prmMeta := &p.ParameterMetadata{Dfv: c.hdbDfv}
 
 	if err := c.pr.IterateParts(func(ph *p.PartHeader) {
 		switch ph.PartKind {
@@ -1432,7 +1430,7 @@ func (c *conn) _readCall(outputFields []*p.ParameterField) (*callResult, []p.Loc
 	rows := &p.RowsAffected{}
 	var ids []p.LocatorID
 	outPrms := &p.OutputParameters{}
-	meta := &p.ResultMetadata{}
+	meta := &p.ResultMetadata{Dfv: c.hdbDfv}
 	resSet := &p.Resultset{}
 	lobReply := &p.WriteLobReply{}
 	var numRow int64