+++ RELEASENOTES.md
@@ -5,6 +5,10 @@ Release Notes
 
 ### Minor revisions
 
+#### v1.1.5
+- added build tag 'edan' to return NULL in case of empty dates for all data format versions
+- removed old build tags from source code
+
 #### v1.1.4
 - performance improvements
 - updated dependencies